+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.0.1-dev-15-feature/DHP-21-surveyjs-format
+- Add an endpoint to retrieve Survey.js json format
+
+## 0.0.1-dev-14-feature/DHP-24-options-url-2
+- Options URLs refactored.
+- Options URL json fields added.
+
+## 0.0.1-dev-13-feature/DHP-22-options-widget-2
+- Textarea filed for items converted to a options field
+
 ## 0.0.1-dev-12-feature/DHP-15-get-survey
 - Get Survey(s) detail endpoint added.
 - Validation rules removed.